common: move cspsettings from web.api to core.common project
