+++ web/ASC.Web.Api/Core/CspSettingsHelper.cs
@@ -29,20 +29,6 @@ using ASC.Web.Files.Classes;
 
 namespace ASC.Web.Api.Core;
 
-public class CspSettings : ISettings<CspSettings>
-{
-    [JsonIgnore]
-    public Guid ID => new Guid("27504162-16FF-405F-8530-1537B0F2B89D");
-
-    public IEnumerable<string> Domains { get; set; }
-    public bool SetDefaultIfEmpty { get; set; }
-
-    public CspSettings GetDefault()
-    {
-        return new CspSettings();
-    }
-}
-
 [Scope]
 public class CspSettingsHelper
 {
